third_party/WebKit/LayoutTests/fast/media/mq-aspect-ratio-overflow.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  @media (max-aspect-ratio: 4294967295/1) { #t1 { color:green } }
  @media (max-aspect-ratio: 4294967296/1) { #t2 { color:green } }
</style>
<div id="t1">This text should be green.</div>
<div id="t2">This text should be green.</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(t1).color, "rgb(0, 128, 0)", "#t1 should be green.");
  }, "Aspect ratio with numerator = 2^32-1.");
  test(() => {
    assert_equals(getComputedStyle(t2).color, "rgb(0, 128, 0)", "#t2 should be green.");
  }, "Aspect ratio with numerator = 2^32.");
</script>
